# apps/rhapi/src/resources/swagger.json
{
   "openapi":"3.0.0",
   "info":{
      "title":"API RH",
      "version":"1.0.0",
      "description": "Application pour apprentissage des outils et méthodes de test.\n\nDans les exemples ci dessous, l'application répond sur http://{server}:{port}"
   },
   "tags": [
    {
       "name": "API - Métier",
       "description": "Expose les routes métier"
     },
    {
       "name": "API - Admin",
       "description": "Expose les routes d'administration"
     }
   ],
   "paths":{
      "/api/employees":{
         "get":{
            "summary":"Récupérer tous les employés",
            "description":"Retourne la liste complète des employés.",
            "tags": ["API - Métier"],
            "responses":{
               "200":{
                  "description":"Liste des employés",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/Employee"
                           }
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/rechercher":{
         "post":{
            "summary":"Recherche avancée (fulltext ou multi-critères)",
            "description":"Recherche sur tous les champs (fulltext) ou par critères spécifiques.",
            "tags": ["API - Métier"],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "type":"object",
                        "properties":{
                           "mode":{
                              "type":"string",
                              "enum": ["fulltext", "detailed"],
                              "description":"Mode de recherche obligatoire: 'fulltext' pour recherche globale, 'detailed' pour multi-critères"
                           },
                           "search":{
                              "oneOf": [
                                 { "type": "string", "description": "Texte pour recherche fulltext" },
                                 { "type": "object", "properties": {
                                    "id": { "type": "string" },
                                    "name": { "type": "string" },
                                    "lastname": { "type": "string" },
                                    "salary": { "type": "string" },
                                    "level": { "type": "string" }
                                 }, "description": "Objets pour recherche multi-critères" }
                              ]
                           }
                        },
                        "required":["mode","search"]
                     },
                     "examples":{
                        "fulltext":{
                           "summary":"Recherche globale",
                           "value":{ "mode":"fulltext", "search":"rené" }
                        },
                        "detailed":{
                           "summary":"Recherche multi-critères",
                           "value":{ "mode":"detailed", "search":{ "name":"Durand", "level":"1" } }
                        },
                        "erreur":{
                           "summary":"Erreur mode absent ou incorrect",
                           "value":{ "search":"Durand" }
                        }
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"Liste des employés trouvés",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/Employee"
                           }
                        }
                     }
                  }
               },
               "400":{
                  "description":"Demande mal formée"
               }
            }
         }
      },
            "/api/employee/{id}":{
               "get":{
                  "summary":"Récupérer un employé par id (matricule)",
                  "description":"Récupérer un employé par son matricule unique.",
                  "tags": ["API - Métier"],
                  "parameters":[
                     {
                        "name":"id",
                        "in":"path",
                        "description":"Matricule de l'employé",
                        "required":true,
                        "schema":{
                           "type":"string"
                        }
                     }
                  ],
                  "responses":{
                     "200":{
                        "description":"Employé trouvé",
                        "content":{
                           "application/json":{
                              "schema":{
                                 "$ref":"#/components/schemas/Employee"
                              }
                           }
                        }
                     },
                     "404":{
                        "description":"Employé non trouvé"
                     }
                  }
               }
            },
      "/api/ajouter":{
         "post":{
            "summary":"Ajouter un employé",
            "description":"Ajouter un employé à la liste des employés",
            "tags": ["API - Métier"],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/Employee"
                     }
                  }
               }
            },
            "responses":{
               "201":{
                  "description":"L'employé a été ajouté avec succès"
               },
               "400":{
                  "description":"L'employé n'a pas été ajouté"
               },
               "409":{
                  "description":"L'employé existe déjà"
               }
            }
         }
      },
      "/api/modifier/{id}":{
         "put":{
            "summary":"Modifier un employé",
            "description":"Modifier les informations d'un employé existant",
            "tags": ["API - Métier"],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"L'ID de l'employé à modifier",
                  "required":true,
                  "schema":{
                     "type":"string"
                  }
               }
            ],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "type":"object",
                        "properties":{
                           "name": { "type": "string" },
                           "lastname": { "type": "string" },
                           "salary": { "type": "string" },
                           "level": { "type": "string" }
                        },
                        "required": ["name", "lastname", "salary", "level"]
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"L'employé a été modifié avec succès"
               },
               "400":{
                  "description":"L'employé n'a pas été modifié"
               },
               "404":{
                "description":"L'employé n'existe pas"
             },
               "409":{
                  "description":"Une erreur est survenue lors de la modification de l'employé"
               }
            }
         }
      },
      "/api/supprimer":{
         "delete":{
            "summary":"Supprimer un employé",
            "description":"Supprimer un employé existant de la liste des employés",
            "tags": ["API - Métier"],
            "parameters":[
               {
                  "name":"id",
                  "in":"query",
                  "description":"L'ID de l'employé à supprimer",
                  "required":true,
                  "schema":{
                     "type":"string"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"L'employé a été supprimé avec succès"
               },
               "400":{
                  "description":"Une erreur est survenue lors de la suppression de l'employé"
               }
            }
         }
      },
      "/api/deleteall":{
          "delete":{
             "summary":"Admin - Supprimer les données",
             "description":"Supprimer les données",
             "tags": ["API - Admin"],
             "parameters":[
                {
                   "name":"authorization",
                   "in":"header",
                   "description":"Admin API token",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"OK"
                },
                "400":{
                   "description":"Bad request"
                },
                "401":{
                   "description":"Token manquant"
                },
                "403":{
                   "description":"Accès interdit, token invalide"
                }
             }
          }
       },
       "/api/datatest":{
          "post":{
             "summary":"Admin - Restaurer les données de test",
             "description":"Restaurer les données de test",
             "tags": ["API - Admin"],
             "parameters":[
                {
                   "name":"authorization",
                   "in":"header",
                   "description":"Admin API token",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "responses":{
                "201":{
                   "description":"OK"
                },
                "400":{
                   "description":"Bad request"
                },
                "401":{
                   "description":"Token manquant"
                },
                "403":{
                   "description":"Accès interdit, token invalide"
                }
             }
          }
       },
       "/metrics":{
          "get":{
             "summary":"Admin - Exporter les métriques Prometheus",
             "description":"Expose les métriques Prometheus pour la supervision. Peut être activé/désactivé dynamiquement.",
             "tags": ["API - Admin"],
             "responses":{
                "200":{
                   "description":"Métriques Prometheus exportées (format texte)"
                },
                "403":{
                   "description":"Endpoint désactivé"
                }
             }
          }
       },
       "/metric":{
          "get":{
             "summary":"Admin - Activer/Désactiver l'exposition des métriques",
             "description":"Active ou désactive dynamiquement l'exposition de la route /metrics via le paramètre enable=true|false.",
             "tags": ["API - Admin"],
             "parameters": [
                {
                  "name": "enable",
                  "in": "query",
                  "description": "true pour activer, false pour désactiver l'exposition de /metrics",
                  "required": true,
                  "schema": {
                    "type": "string",
                    "enum": ["true", "false"]
                  }
                }
             ],
             "responses":{
                "200":{
                   "description":"État mis à jour (metricsEnabled: true/false)"
                },
                "400":{
                   "description":"Paramètre 'enable' requis ou invalide"
                }
             }
          }
       }
   },
   "components":{
      "schemas":{
         "Employee":{
            "type":"object",
            "properties":{
               "id":{
                  "type":"string"
               },
               "name":{
                  "type":"string"
               },
               "lastname":{
                  "type":"string"
               },
               "salary":{
                  "type":"string"
               },
               "level":{
                  "type":"string"
               },
               "time":{
                  "type":"number"
               }
            },
            "required":[
               "id",
               "name",
               "lastname",
               "salary",
               "level",
               "time"
            ]
         }
      }
   }
 }